MSMixVPR/requirements.txt
faiss-cpu==1.7.2
# faiss-gpu==1.7.2
matplotlib==3.5.3
numpy==1.23.4
pandas==1.5.1
Pillow==9.2.0
prettytable==3.5.0
pytorch-lightning==1.8.3
pytorch-metric-learning==1.6.3
scikit-learn==1.2.1
scipy==1.9.3
tensorboard==2.11.0
timm==0.6.12
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.11.1
torchvision==0.14.1
tqdm==4.64.1
# psutil
keras
# fvcore
